// packages/core/src/zod/restaurant.schema.d.ts
export declare const restaurantWithRelationsSchema: z.ZodObject<{
    id: z.ZodString;
    hotelId: z.ZodString;
    createdBy: z.ZodString;
    organizationId: z.ZodString;
    name: z.ZodString;
    description: z.ZodNullable<z.ZodString>;
    brandName: z.ZodNullable<z.ZodString>;
    buffetMetadata: z.ZodNullable<z.ZodString>;
    street: z.ZodString;
    city: z.ZodString;
    state: z.ZodString;
    country: z.ZodString;
    postalCode: z.ZodString;
    latitude: z.ZodNullable<z.ZodCoercedString<unknown>>;
    longitude: z.ZodNullable<z.ZodCoercedString<unknown>>;
    phone: z.ZodNullable<z.ZodString>;
    email: z.ZodNullable<z.ZodString>;
    website: z.ZodNullable<z.ZodString>;
    starRating: z.ZodNullable<z.ZodCoercedString<unknown>>;
    checkInTime: z.ZodNullable<z.ZodString>;
    checkOutTime: z.ZodNullable<z.ZodString>;
    cuisineType: z.ZodNullable<z.ZodString>;
    dressCode: z.ZodNullable<z.ZodString>;
    totalSeats: z.ZodNullable<z.ZodNumber>;
    allocatedSeats: z.ZodNullable<z.ZodNumber>;
    breakfastPrice: z.ZodNullable<z.ZodCoercedString<unknown>>;
    lunchPrice: z.ZodNullable<z.ZodCoercedString<unknown>>;
    dinnerPrice: z.ZodNullable<z.ZodCoercedString<unknown>>;
    buffetPrice: z.ZodNullable<z.ZodCoercedString<unknown>>;
    pricePerSeat: z.ZodNullable<z.ZodCoercedString<unknown>>;
    customPrices: z.ZodDefault<z.ZodOptional<z.ZodArray<z.ZodObject<{
        label: z.ZodString;
        price: z.ZodCoercedNumber<unknown>;
    }, z.core.$strip>>>>;
    menuUrl: z.ZodNullable<z.ZodString>;
    status: z.ZodEnum<{
        pending_approval: "pending_approval";
        active: "active";
        inactive: "inactive";
        under_maintenance: "under_maintenance";
    }>;
    createdAt: z.ZodCoercedString<unknown>;
    updatedAt: z.ZodNullable<z.ZodCoercedString<unknown>>;
    images: z.ZodOptional<z.ZodArray<z.ZodObject<{
        id: z.ZodString;
        restaurantId: z.ZodString;
        imageUrl: z.ZodString;
        altText: z.ZodNullable<z.ZodString>;
        displayOrder: z.ZodNullable<z.ZodNumber>;
        isThumbnail: z.ZodNullable<z.ZodBoolean>;
        createdAt: z.ZodCoercedString<unknown>;
        updatedAt: z.ZodNullable<z.ZodCoercedString<unknown>>;
    }, z.core.$strip>>>;
}, z.core.$strip>;